Scan ports.
First commit related to scanning ports. This includes an actual scan of
the ports on a system with success and failure callbacks. Also, this
will fire a complete callback which happens at the end of the scan.

There are a number of options, undocumented at the moment.

There are a handful of tests but nearly extensive.

Signed-off-by: Nick Campbell <nicholas.j.campbell@gmail.com>